+++ app/src/main/java/org/qtum/mromanovsky/qtum/ui/fragment/SetTokenParametersFragment/SetTokenParametersFragmentInteractorImpl.java
@@ -1,8 +1,6 @@
 package org.qtum.mromanovsky.qtum.ui.fragment.SetTokenParametersFragment;
 
 
-import android.content.Context;
-
 import org.qtum.mromanovsky.qtum.datastorage.QtumToken;
 
 class SetTokenParametersFragmentInteractorImpl implements SetTokenParametersFragmentInteractor{
@@ -11,7 +9,6 @@ class SetTokenParametersFragmentInteractorImpl implements SetTokenParametersFrag
 
     }
 
-
     @Override
     public void setInitialSupply(String initialSupply) {
         QtumToken.getInstance().setInitialSupply(Long.parseLong(initialSupply));
@@ -19,7 +16,7 @@ class SetTokenParametersFragmentInteractorImpl implements SetTokenParametersFrag
 
     @Override
     public void setDecimalUnits(String decimalUnits) {
-        QtumToken.getInstance().setDecimalUnits(Long.parseLong(decimalUnits));
+        QtumToken.getInstance().setDecimalUnits(Integer.parseInt(decimalUnits));
     }
 
     @Override
